+++ android/app/src/main/java/com/example/aida/MainActivity.kt
@@ -4,17 +4,7 @@ import android.content.Context
 import android.os.Bundle
 import androidx.activity.ComponentActivity
 import androidx.activity.compose.setContent
-import androidx.compose.foundation.Image
-import androidx.compose.foundation.background
-import androidx.compose.foundation.clickable
-import androidx.compose.foundation.layout.Arrangement
-import androidx.compose.foundation.layout.Column
-import androidx.compose.foundation.layout.Row
-import androidx.compose.foundation.layout.Spacer
-import androidx.compose.foundation.layout.height
-import androidx.compose.foundation.layout.offset
 import androidx.compose.foundation.layout.padding
-import androidx.compose.foundation.layout.width
 import androidx.compose.material.icons.Icons
 import androidx.compose.material.icons.automirrored.filled.MenuBook
 import androidx.compose.material.icons.filled.Home
@@ -29,30 +19,26 @@ import androidx.compose.material3.NavigationDrawerItem
 import androidx.compose.material3.Surface
 import androidx.compose.material3.Text
 import androidx.compose.material3.rememberDrawerState
-import androidx.compose.runtime.Composable
-import androidx.compose.runtime.LaunchedEffect
-import androidx.compose.runtime.collectAsState
 import androidx.compose.runtime.getValue
 import androidx.compose.runtime.mutableIntStateOf
 import androidx.compose.runtime.mutableStateOf
 import androidx.compose.runtime.remember
 import androidx.compose.runtime.rememberCoroutineScope
 import androidx.compose.runtime.setValue
-import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
-import androidx.compose.ui.draw.scale
 import androidx.compose.ui.platform.LocalConfiguration
 import androidx.compose.ui.platform.LocalContext
-import androidx.compose.ui.res.painterResource
-import androidx.compose.ui.unit.Dp
 import androidx.compose.ui.unit.dp
-import androidx.compose.ui.unit.sp
 import androidx.datastore.core.DataStore
 import androidx.datastore.preferences.core.Preferences
 import androidx.datastore.preferences.preferencesDataStore
 import androidx.lifecycle.viewmodel.compose.viewModel
+import com.example.aida.pages.CameraPage
+import com.example.aida.pages.ConfigurationPage
+import com.example.aida.ui.composables.TopBar
 import com.example.aida.ui.theme.AIDATheme
-import com.example.aida.ui.theme.TopBarColor
+import com.example.aida.viewmodels.MainViewModel
+import com.example.aida.viewmodels.MainViewModelFactory
 import kotlinx.coroutines.launch
 
 /**
@@ -60,9 +46,7 @@ import kotlinx.coroutines.launch
  * and different widgets as well as the main application. The main switches
  * between a camera/lidar feed and configuration page
  *
- * TODO: Fetch real data for camera feed
  * TODO: Fetch real data for lidar feed
- * TODO: Use classes for camera and configuration page?
  * TODO: Decide on what widgets in top bar should do
  *
  * @author Elias
@@ -77,64 +61,6 @@ class MainActivity : ComponentActivity() {
             val context = LocalContext.current
             val dataStore = context.dataStore
             val viewModel: MainViewModel = viewModel(factory = MainViewModelFactory(dataStore))
-
-            val ipAddress by viewModel.ipAddress.collectAsState()
-            val port by viewModel.port.collectAsState()
-
-            /*
-            var cameraFeedConnectionStage by remember { mutableStateOf(ConnectionStages.CONNECTING) }
-            var lidarConnectionStage by remember { mutableStateOf(ConnectionStages.CONNECTING) }
-            var sttConnectionStage by remember { mutableStateOf(ConnectionStages.CONNECTING) }
-            var imageBitmap by remember { mutableStateOf<ImageBitmap?>(null) }
-
-            suspend fun connectToAIDA() {
-                cameraFeedConnectionStage = ConnectionStages.CONNECTING
-                lidarConnectionStage = ConnectionStages.CONNECTING
-                sttConnectionStage = ConnectionStages.CONNECTING
-
-                withContext(Dispatchers.IO) {
-                    try {
-                        val videoClient = VideoClient(
-                            ip = ipAddress,
-                            port = port,
-                            timeToTimeout = 10000
-                        )
-                        videoClient.sendStartCamera()
-                        videoClient.sendGetVideo()
-
-                        cameraFeedConnectionStage = ConnectionStages.CONNECTION_SUCCEEDED
-
-                        while (true) {
-                            val byteArray = videoClient.fetch()
-                            imageBitmap =
-                                BitmapFactory.decodeByteArray(byteArray, 0, byteArray.size)
-                                    ?.asImageBitmap()
-                        }
-                    } catch (e: Exception) {
-                        println("Network Custom Exception: $e")
-                        cameraFeedConnectionStage = ConnectionStages.CONNECTION_FAILED
-                    }
-                    try {
-                        speechViewModel.initializeSTTClient(
-                            ip = ipAddress,
-                            port = port,
-                            timeToTimeout = 5000
-                        )
-                        sttConnectionStage = ConnectionStages.CONNECTION_SUCCEEDED
-                    } catch (e: Exception) {
-                        // TODO Remove lidar line below, only used for testing
-                        lidarConnectionStage = ConnectionStages.CONNECTION_FAILED
-                        sttConnectionStage = ConnectionStages.CONNECTION_FAILED
-                        print("Speech Exception: $e")
-                    }
-                    // TODO Implement lidar fetch
-                }
-            }
-             */
-
-            LaunchedEffect(Unit) {
-                viewModel.connectToAIDA()
-            }
 
             AIDATheme {
                 var topBarTitle by remember { mutableStateOf("AIDA Remote Control Beta") }
@@ -198,6 +124,7 @@ class MainActivity : ComponentActivity() {
                         val configuration = LocalConfiguration.current
                         val screenHeight = configuration.screenHeightDp.dp
                         val screenWidth = configuration.screenWidthDp.dp
+
                         // Dynamically set bar height depending on phone/tablet
                         val barHeight = if (screenHeight / 8 < 50.dp) screenHeight / 6 else 50.dp
 
@@ -215,44 +142,20 @@ class MainActivity : ComponentActivity() {
                             topBarTitle = topBarTitle
                         )
 
-                        val configurationCoroutineScope = rememberCoroutineScope()
-
+                        // Main logic when switching between pages
                         when (state) {
                             0 -> CameraPage(
                                 screenHeight,
                                 barHeight,
                                 screenWidth,
-                                viewModel,
-                                /*
-                                viewModel.imageBitmap.collectAsState().value,
-                                viewModel.cameraFeedConnectionStage.collectAsState().value,
-                                viewModel.sttConnectionStage.collectAsState().value,
-                                viewModel.lidarConnectionStage.collectAsState().value,
-                                ipAddress,
-                                port
-                                 */
+                                viewModel
                             )
 
                             1 -> ConfigurationPage(
                                 barHeight,
-                                ipAddress = ipAddress,
-                                onIpAddressChange = { newIp ->
-                                    configurationCoroutineScope.launch {
-                                        viewModel.updateIpAddress(newIp)
-                                    }
-                                },
-                                port = port,
-                                onPortChange = { newPort ->
-                                    configurationCoroutineScope.launch {
-                                        viewModel.updatePort(newPort)
-                                    }
-                                },
+                                viewModel,
                                 onButtonPress = {
                                     state = 0
-
-                                    configurationCoroutineScope.launch {
-                                        viewModel.connectToAIDA()
-                                    }
                                     topBarTitle = "AIDA Remote Control Beta"
                                 }
                             )
@@ -260,125 +163,6 @@ class MainActivity : ComponentActivity() {
                     }
                 }
             }
-        }
-    }
-}
-
-/**
- * Function for the top bar, contains both UI and logic, however some
- * widgets need to be decided on what they should do.
- *
- * @param onMenuClicked records if the menu button has been pressed
- * @param onCameraClicked records if the camera on button has been pressed
- * @param barHeight used to determine the size of the top bar
- * @param topBarTitle used when want to dynamically change the title
- * @author Elias
- */
-@Composable
-fun TopBar(
-    onMenuClicked: () -> Unit,
-    onCameraClicked: () -> Unit,
-    screenWidth: Dp,
-    barHeight: Dp,
-    topBarTitle: String,
-) {
-    val barPadding = 15.dp
-
-    Row(
-        modifier = Modifier
-            .height(barHeight)
-            .width(screenWidth)
-            .background(TopBarColor.copy(alpha = 0.6f))
-            .padding(start = barPadding, end = barPadding),
-        horizontalArrangement = Arrangement.Absolute.SpaceEvenly,
-        verticalAlignment = Alignment.CenterVertically
-    ) {
-        Row(
-            modifier = Modifier
-                .weight(1f)
-        ) {
-            Image(
-                painter = painterResource(id = R.drawable.configuration_button),
-                contentDescription = "configuration",
-                Modifier
-                    .clickable(onClick = onMenuClicked)
-                    .scale(1.2f)
-            )
-        }
-        Row(
-            modifier = Modifier.weight(1f),
-            horizontalArrangement = Arrangement.Center
-        ) {
-            Text(
-                text = topBarTitle,
-                fontSize = 16.sp
-            )
-        }
-        Row(
-            modifier = Modifier.weight(1f),
-            horizontalArrangement = Arrangement.End,
-            verticalAlignment = Alignment.CenterVertically
-        ) {
-            var cameraPress by remember { mutableStateOf("on") }
-            var speakerPress by remember { mutableStateOf("on") }
-
-            Spacer(Modifier.weight(5f))
-
-            Column(
-                horizontalAlignment = Alignment.CenterHorizontally, // Aligns the contents of the column to the center horizontally
-                modifier = Modifier
-                    .clickable(onClick = {
-                        cameraPress = if (cameraPress == "on") "off" else "on"
-                        onCameraClicked()
-                    }
-                    )
-                    .padding(8.dp) // Adds padding around the column
-            ) {
-                Image(
-                    painter = painterResource(id = R.drawable.camera_button),
-                    contentDescription = "camera",
-                    Modifier
-                        .scale(1.2f)
-                )
-                Text(
-                    text = cameraPress,
-                    Modifier.offset(y = (-2).dp)
-                )
-            }
-            Spacer(Modifier.weight(1f))
-
-            Column(
-                horizontalAlignment = Alignment.CenterHorizontally,
-                modifier = Modifier
-                    .clickable(onClick = {
-                        speakerPress = if (speakerPress == "on") "off" else "on"
-
-                    }
-                    )
-                    .padding(8.dp)
-            ) {
-                Image(
-                    painter = painterResource(id = R.drawable.volume_button),
-                    contentDescription = "camera",
-                    Modifier
-                        .scale(1.2f)
-                )
-                Text(
-                    text = speakerPress,
-                    Modifier.offset(y = (-2).dp)
-                )
-            }
-            Spacer(Modifier.weight(1f))
-            Text(
-                text = "Battery: " + 69 + "%",
-                fontSize = 16.sp
-            )
-            Spacer(Modifier.weight(1f))
-
-            Text(
-                text = "Lag: " + 12 + "ms",
-                fontSize = 16.sp
-            )
         }
     }
 }
