+++ android/app/src/main/java/com/example/aida/enums/ConnectionStages.kt
@@ -1,4 +1,4 @@
-package com.example.aida
+package com.example.aida.enums
 
 enum class ConnectionStages {
     CONNECTING,
